добавляет класс смарт и фиксит баг
